footer and other things

+++ src/api/models.py
@@ -36,7 +36,7 @@ class Product(db.Model):
     name = db.Column(db.String(120), nullable=False)
     stock = db.Column(db.Integer)
     type = db.Column(db.String(120))
-    price = db.Column(db.Integer)
+    price = db.Column(db.FLOAT())
     unit = db.Column(db.String(120))
     location = db.Column(db.String(120))
     description = db.Column(db.String(120))
